improve ui & add github banner in navbar
