+++ app/layout.tsx
@@ -6,7 +6,10 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
+    <html
+      lang="en"
+      className="scrollbar-track-transparent scrollbar-thumb-foreground/10"
+    >
       <body>{children}</body>
     </html>
   );
